Move MTU detection to separate module
